requirements.txt
# Core LLM / RAG stack
openai>=1.50.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.6.0
tiktoken>=0.8.0

# Document handling
pymupdf>=1.24.0
python-docx>=1.1.0
jinja2>=3.1.0

# Data / validation
pydantic>=2.6.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0

# Utilities
tqdm>=4.66.0

# Web UI
streamlit>=1.40.0
